Moved celery test files
