+++ music_feed/celery_test/__init__.py
@@ -1,0 +1,2 @@
+
+from .upload_celery_tasks import update_all_channels
