--- music_feed/youtube_data/__init__.py
@@ -1,2 +1,0 @@
-
-from .get_uploads import update_all_channels
